### 2022-08-20 #4685 @editorialbot
[PRE REVIEW]: GSAreport: Easy to Use Global Sensitivity Reporting
**Submitting author:** @Basvanstein (<a href="http://orcid.org/0000-0002-0013-7969">Bas van Stein</a>)
**Repository:** https://github.com/Basvanstein/GSAreport
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
