### 2022-08-25 9518b73
Update README.md
Renamed the docker image to the automically deployed one.
