### 2022-09-16 Basvanstein/GSAreport#14 @xtruan
Include sample CSV files in repository
The example documentation refers to several CSV files in order to conduct a test run of the application. However, there are no samples in the repository. It would be helpful to include some samples so an end user could try out the software without having their own data.

From README.md:
- x.csv, y.csv #optional, in case you use an existing design of experiments
- x_sobol.csv, y_sobol.csv
- x_morris.csv, y_morris.csv
- x_lhs.csv, y_lhs.csv
