### 2022-09-16 Basvanstein/GSAreport#15 @xtruan
Add community guidelines
JOSS requires some community guidelines. Add guidelines for third parties wishing to:
1) Contribute to the software
2) Report issues or problems with the software
3) Seek support
