### 2022-09-16 Basvanstein/GSAreport#15 @xtruan
I realized that there are some community/contribution guidelines on the GitHub Pages site. Can you provide a link to this documentation in the repo metadata or README.md file?

https://basvanstein.github.io/GSAreport/1.0/
