### 2022-09-16 #4721 @xtruan
First thank you for submitting this interesting project! I don't have much domain expertise in this area, however this seems like a useful tool.

I have created a few issues in the repository: https://github.com/Basvanstein/GSAreport/issues/13, https://github.com/Basvanstein/GSAreport/issues/14, https://github.com/Basvanstein/GSAreport/issues/15.

I would like to be able to build the container and test locally with some provided sample data to verify the performance of the tool.

In addition, I have a few other points:
1) I want to understand the contributions of @elenaraponi as it appears the vast majority of the code was committed by @Basvanstein. Can you provide some context as to your roles?
2) Can you add some context as to the "State of the field" in the paper? It would be helpful to know how this tool expands/improves on the SALib and SHAP libraries mentioned in the references.
