### 2022-09-17 Basvanstein/GSAreport#13 @Basvanstein
Dear @xtruan , thank you for reviewing and also trying the local build.
I used python 3.9 which is I think why your build fails as 3.10 was not tested and likely not supported by all packages.
Would you like me to update the dependencies to python 3.10 or can you run it with 3.9?
Also it helps me to know what OS you used, I am using mac to build the dockerfile, on windows it might behave slightly different.
Instead of building the docker file locally you could also try the github action to build it, this is how we currently do it. Github then immediately stores and serves the docker image as a release.
