### 2022-09-17 Basvanstein/GSAreport#14 @Basvanstein
Good point will upload csv files.
Will also add the following command to the readme

    docker run -v `pwd`/output:/output -v `pwd`/data:/data ghcr.io/basvanstein/gsareport:main --demo

The demo command will generate sample data as well.
