### 2022-09-18 2adb14e
updated requirements and added sample csvs
Updated readme to include documentation link Resolves #15

Added sample csv files to resolve #14

Updated requirements for python 3.10 Resolves #13
