+++ src/requirements.txt
@@ -20,7 +20,7 @@ debugpy==1.5.1
 decorator==5.1.0
 defusedxml==0.7.1
 dill==0.3.4
-distlib==0.3.3
+distlib
 distro==1.6.0
 entrypoints==0.3
 fonttools==4.28.5
@@ -41,7 +41,7 @@ jupyter-console==6.4.0
 jupyter-core==4.9.1
 jupyterlab-pygments==0.1.2
 kiwisolver==1.3.2
-llvmlite==0.38.1
+llvmlite==0.39.1
 MarkupSafe==2.0.1
 matplotlib==3.5.1
 matplotlib-inline==0.1.3
@@ -54,8 +54,8 @@ nbconvert==6.1.0
 nbformat==5.1.3
 nest-asyncio==1.5.4
 notebook==6.4.4
-numba==0.55.2
-numpy==1.21.5
+numba==0.56.2
+numpy==1.23.3
 ordered-set==4.0.2
 packaging==21.3
 pandas==1.3.5
@@ -74,15 +74,15 @@ progress==1.6
 prometheus-client==0.11.0
 prompt-toolkit==3.0.24
 ptyprocess==0.7.0
-pycairo==1.20.1
+pycairo
 pycparser==2.21
 Pygments==2.11.1
-PyGObject==3.42.0
+PyGObject
 pyOpenSSL==21.0.0
 pyparsing==2.4.7
 pyrsistent==0.18.0
 python-dateutil==2.8.2
-pytz==2021.3
+pytz
 PyYAML==6.0
 pyzmq==22.2.1
 requests==2.26.0
@@ -90,7 +90,7 @@ resolvelib==0.5.5
 retrying==1.3.3
 SALib==1.4.5
 scikit-learn==1.0.2
-scipy==1.7.3
+scipy==1.9.1
 Send2Trash==1.8.0
 shap==0.40.0
 six==1.16.0
