### 2022-09-18 #4721 @Basvanstein
Thank you @xtruan for the review so far.
I have resolved the 3 reported issues in the last commit.
The local build was failing because the base docker image was updated, I have updated the dependencies such that it works correctly again.

For your other questions:

1. @elenaraponi contributed by testing, evaluating the user experience and helping with the mathematics of some of the methods. It is true that she did not commit a lot of code, but I could not have done it without her.
2. GSAreport unifies the methods of SHAP and SALib in one easy to use package. Each method in the original SALib package requires a different sampling scheme and with different sample sizes, in GSAreport we try to streamline this and to allow for any arbitrary sample scheme and size, making it a lot easier to use. In addition we use empirically found rules of thumb on which GSA methods work well for a given sample size and dimensionality as not all methods provide meaningful results when the dimensionality is high and the sample size low. For additional details see also our (just released) preprint paper on [techrxiv](https://www.techrxiv.org/articles/preprint/A_Comparison_of_Global_Sensitivity_Analysis_Methods_for_Explainable_AI_with_an_Application_in_Genomic_Prediction/20528292)

Let me know if you encounter any other issues.
