### 2022-09-20 Basvanstein/GSAreport#16 @rreinecke
JOSS Review - References on sampling schme adapation
Reading through the JOSS paper and the Readme of your project I find myself wondering what exactly the added value of your application is. I am not saying it is not there I just don't quite understand why I would not just be using SAlib directly. Also I think you should point out other toolboxes like SAFE.

To be more precise: Please add a little bit more description of what the user explicitly gains from using this software. Be a bit more explicit on how it fits into the world of existing toolboxes out there.

Furthermore, you state that "each method in the original SALib package requires a different sampling scheme and different sample size". Well the sample size is related to convergence and the scheme very much to the underlying methods and has nothing to do with SAlib. Please elaborate what your your tool does exactly. For me this is currently rather confusing.
