### 2022-09-20 Basvanstein/GSAreport#16 @rreinecke
https://github.com/openjournals/joss-reviews/issues/4721
