### 2022-09-20 #4721 @rreinecke
Dear @Basvanstein adding to the already stated remarks I opened an issue https://github.com/Basvanstein/GSAreport/issues/16 concerning additional notes on the added value of the software.
