### 2022-09-20 Basvanstein/GSAreport#16 @Basvanstein
Dear @rreinecke , happy to explain it more elaborately of course, but I am not sure how to make this concise in the paper.md.

The GSAreport application is different from using SALib directly in a number of ways:

1. The GSAreport application does not require python knowledge (though can be used as python package as well).
2. The GSAreport application uses a surrogate model when only 1 specific sampling scheme is present (for example for expensive industrial applications) to allow other GSA methods to work properly.
3. Many GSA methods require a sampling scheme with a specific number of evaluations in terms of the dimensions (for example x*(d+1) or x*(d+2), this makes it hard to run multiple GSA methods if you are not familiar with these details, GSAreport tries to take this problem away from the user by working around these requirements.
4. The GSAreport application selects a suitable subset of GSA methods based on empirically found rules of thumb, for specifics about this selection I would like to refer to our paper: https://www.techrxiv.org/articles/preprint/A_Comparison_of_Global_Sensitivity_Analysis_Methods_for_Explainable_AI_with_an_Application_in_Genomic_Prediction/20528292
5. The GSAreport application bundles these methods in a readable report with interactive plots showing both first and second order sensitivities, including explanations and references to the different methods.
6. GSAreport also uses graph-tool to visualise first and second order Sobol indices in an intuitive manner.

SAFE is indeed a comparable toolbox but is for Matlab and I have unfortunately no experience with the toolbox and as such I find it difficult to compare it, I can of course add it as a relevant link.
GSAreport is meant for making GSA methods as easy as possible to use and to advice the end-users which methods work well in which situation.
I hope this clarifies the contribution. Do you have any suggestion as to how to incorporate this better in the paper.md?
