### 2022-09-20 Basvanstein/GSAreport#16 @rreinecke
Thank you @Basvanstein for the additional explanation! It is much clearer to me now. Why not just integrate this text into the paper? You could add it as subsection to the statement of need.
I think this statement is quite important and I am not sure this message gets through in the current paper version: "GSAreport is meant for making GSA methods as easy as possible to use and to advice the end-users which methods work well in which situation." Please think about revising the text and making it more concise.

SAFE is actually also available in python. And they also provide "workflows" that users can follow to apply the methods in a meaningful way. I agree that this is more applied than your tool but maybe still noteworthy.

Overall I like to add that I really like your project and think this is a valuable addition to the SA community!
