+++ paper.md
@@ -35,7 +35,15 @@ and somewhat complex to implement GSA methods and sampling techniques.
 
 `GSAreport` was designed to be used by both ML researchers as well as by domain experts in industry to validate and inspect different machine learning models and simulators and gain better insights into their complex industrial processes.
 
-`GSAreport` unifies methods of the SHAP and SALib packages in one easy to use application with an interactive graphical report. Each method in the original SALib package requires a different sampling scheme and different sample size, in GSAreport we streamline this to allow for any arbitrary sample scheme and size. In addition we use empirically found rules of thumb to automatically select GSA methods that work well for a given sample size and dimensionality.
+`GSAreport` unifies methods of the SHAP and SALib packages in one easy to use application with an interactive graphical report. 
+The GSAreport application is different from using SALib directly in a number of ways:
+
+* `GSAreport` does not require python knowledge (though can be used as python package as well).
+* `GSAreport` uses a surrogate model when only 1 specific sampling scheme is present (for example for expensive industrial applications) to allow other GSA methods to work properly.
+* Many GSA methods require a sampling scheme with a specific number of evaluations in terms of the dimensions (for example x*(d+1) or x*(d+2), this makes it hard to run multiple GSA methods if you are not familiar with these details, GSAreport tries to take this problem away from the user by working around these requirements.
+* `GSAreport` selects a suitable subset of GSA methods based on empirically found rules of thumb.
+
+`GSAreport` bundles the GSA methods in a readable report with interactive plots showing both first and second order sensitivities, including explanations and references to the different methods. The application is meant for making GSA methods as easy as possible to use and to advice the end-users about which methods work well under which conditions.
 
 # References
 
