+++ paper.md
@@ -52,6 +52,8 @@ The following open source software libraries are used (and extended) within `GSA
 - SALib [SALibHerman2017], a sensitivity analysis library for Python containing different SA methods and sampling schemes.  
 - SHAP [shapNIPS2017_7062], a local SA method for explainable AI.
 
+A related toolbox originally implemented for Matlab and also available for Python and R is the SAFE toolbox [pianosi2015matlab].
+
 # Acknowledgements
 
 This work is partly funded by the Dutch Research Council (NWO) as part of the XAIPre project (with project number 19455).
