### 2022-09-20 Basvanstein/GSAreport#16 @Basvanstein
Thank you very much for the nice comment.
I have updated the paper, for some reason I had in my mind it could be 1 page maximum, but I realise now that that is not the case. If you think the additions to the paper resolve the issue feel free to close it.
