### 2022-09-28 Basvanstein/GSAreport#16 @rreinecke
Great! I saw that you have updated the md file in your paper branch. This seems not to have changed the paper generated by the editorial bot. Anyways I am closing this. @Basvanstein Please make sure your changes make it into the final paper.
