### 2022-09-28 Basvanstein/GSAreport#18 @rreinecke
JOSS review: notes on example usage
- The demo mode on the docker image doesn't seem to work
- Please provide the files necessary to run the example at https://basvanstein.github.io/GSAreport/1.0/usecases It is nicely described but providing the first json file would make it much easier to get started.
- I am a bit confused by the example: you state that the user does not necessary need any python knowledge. But for running the SA in the quickstart you really need to know what you are doing even if it is a simple example. So GSA report is not running the SA for you? It is simplifying the sampling and the analysis but it is not running the experiment. I think that should be clarified in the project and paper description.
- the last line in the quickstart as improper line breaks - a simple copy and paste does not work
- After running the Qucikstart it provides me with the output: "Created a graph with 3 vertices and 1 edges.
Vertices are the top 5 ST values greater than 0.01.
Only S2 values (edges) greater than 0.0 are included." ok well, what does that tell me? The output folder is empty. How do I get the graphical report?
