### 2022-09-28 #4721 @rreinecke
@Basvanstein the docker description is working flawless but I have a couple remarks on the example use case description:
Basvanstein/GSAreport/issues/18
Other remarks:
- Also I was not able to find any automated tests of your application
- Where can I find a documentation of the API? What if I would like to add support for a different SA method not support by SAlib. Or a different sampling scheme. Where would I start as a contributer?
