### 2022-09-28 Basvanstein/GSAreport#18 @Basvanstein
> * The demo mode on the docker image doesn't seem to work

What did you try exactly? It is important that when running Docker we also create a volume and specify the output path such that we see the generated files.
For example the following command:

```
#on linux:
docker run --rm -v `pwd`/output:/output ghcr.io/basvanstein/gsareport:main --demo -o /output

#on windows:
docker run --rm -v $pwd/output:/output ghcr.io/basvanstein/gsareport:main --demo -o /output
```

Runs the docker image with the "output" directory (of the current working directory) as a volume linked to the docker "/output" folder. The `-o /output` tells the program to store the resulting report in the output directory.
After this run you should see an html report in the output directory appear.

> * Please provide the files necessary to run the example at https://basvanstein.github.io/GSAreport/1.0/usecases It is nicely described but providing the first json file would make it much easier to get started.

The example files including the `problem.json` can be found in the `data` directory.

> * I am a bit confused by the example: you state that the user does not necessary need any python knowledge. But for running the SA in the quickstart you really need to know what you are doing even if it is a simple example. So GSA report is not running the SA for you? It is simplifying the sampling and the analysis but it is not running the experiment. I think that should be clarified in the project and paper description.

GSAreport is not running the evaluation (what you call experiment) of the real function / model you want to analyse. This is on purpose to make it flexible as the thing you want to analyse could be an R model, python ML model, executable etc. etc.
GSAreport provides 2 of the 3 steps, generating the required sample schemes (if you do not use a custom scheme) and running the SA algorithms to generate a report. I thought it was well described but I can add some sentences to put more emphasis on this aspect.

> * the last line in the quickstart as improper line breaks - a simple copy and paste does not work

Will check, which terminal do you use? It was written for Linux/mac and works slightly different for powershell for example (I can add a tab for windows as well)

> * After running the Qucikstart it provides me with the output: "Created a graph with 3 vertices and 1 edges.
>   Vertices are the top 5 ST values greater than 0.01.
>   Only S2 values (edges) greater than 0.0 are included." ok well, what does that tell me? The output folder is empty. How do I get the graphical report?

If you use a volume and set the output parameter of the program correctly it should generate a report inside the folder. If it does not appear it is likely that it generated the report in a folder you do not have a docker volume to. This could be related to your terminal environment as well. For example the last command from the quickstart (which you can run on the sample data as well) is for windows powershell:

```
docker run --rm -v $pwd/output:/output -v $pwd/data:/data ghcr.io/basvanstein/gsareport:main -p /data/problem.json -d /data -o /output
```

Hope this helps, let me know what terminal/OS you use that way I can provide more detailed instructions. I will also try to include these differences in the Documentation.
