### 2022-09-29 #4721 @rreinecke
> @rreinecke I have replied on the points of issue 18, for some of these points I need a bit more information to help you out. Regarding your other points:
>
>     * There are indeed no automated tests, the demo mode works as a test and SAlib itself has a large coverage of unit tests. Because it is dockerized and generating an html file rather than a number etc, I did not write specific unit tests for the GSAreport function and I don't think it adds much.
>
>     * API documentation is on the same website as the quick start (API reference on top menu): https://basvanstein.github.io/GSAreport/1.0/gsareport/gsareport/

Thank you for your swift responses. I have to disagree on the unit tests. Your code does not only simply generate a html file but also trains a model and processes complex input and output. I would advice to have unit tests for that.
