### 2022-09-29 Basvanstein/GSAreport#18 @rreinecke
> docker run --rm -v `pwd`/output:/output ghcr.io/basvanstein/gsareport:main --demo -o /output

The command works. But the output folder is empty. Also as a Linux user I expect "/output" to write something to root/output. It would be better to just write "output" without the "/".

> The example files including the problem.json can be found in the data directory.
Then point the user to that file in the documentation. How should I know that?

> GSAreport is not running the evaluation (what you call experiment) of the real function / model you want to analyse.
Maybe you could add a flow chart diagram to the documentation to make clear what the typical workflow is. And yes please state it clearly in the paper.

> Will check, which terminal do you use? It was written for Linux/mac and works slightly different for powershell for example (I can add a tab for windows as well)
I am using the standard Gnome terminal. All other commands did work with copy paste except for this one.

> windows powershell:
This then likely explains a lot ... could you please provide me with the correct command to that I can test that on a proper OS.
Please also state that in the user guide.
