### 2022-09-29 #4721 @Basvanstein
Ok, I can add unit tests for those functions that do not generate files. It will take a bit of time though.
