### 2022-09-29 Basvanstein/GSAreport#18 @Basvanstein
> > docker run --rm -v `pwd`/output:/output ghcr.io/basvanstein/gsareport:main --demo -o /output
>
> The command works. But the output folder is empty. Also as a Linux user I expect "/output" to write something to root/output. It would be better to just write "output" without the "/".

Sorry but this is actually related to Docker knowledge and not the tool, it definitely points to the output folder in the root, but the root is the root of the Docker container which is linked via a volume to the output folder of the current working directory.
Hard to say why the output folder is empty in your case since I tested this on 3 different machines and for all 3 it worked, Ubuntu, Mac and Windows. Perhaps you did not give Docker permission to access your drive? I believe this issue is Docker related and not related to the actual application we propose. Perhaps instead you can try to run the python library instead which is easy to install on linux.
>
> > The example files including the problem.json can be found in the data directory.
>
> Then point the user to that file in the documentation. How should I know that?

It states in the documentation "To start, you always need to provide the program with a problem definition. This definition can be supplied as json file, see also data/problem.json"
Perhaps you read over it, I can make it more clear by repeating it in the data section.
>
> > GSAreport is not running the evaluation (what you call experiment) of the real function / model you want to analyse.
>
> Maybe you could add a flow chart diagram to the documentation to make clear what the typical workflow is. And yes please state it clearly in the paper.

Good idea, will do that.
>
> > Will check, which terminal do you use? It was written for Linux/mac and works slightly different for powershell for example (I can add a tab for windows as well)
>
> I am using the standard Gnome terminal. All other commands did work with copy paste except for this one.
>
Thanks will check with Gnome.

> > windows powershell:
>
> This then likely explains a lot ... could you please provide me with the correct command to that I can test that on a proper OS. Please also state that in the user guide.

For Linux that should be
```
docker run --rm -v `pwd`/output:/output -v `pwd`/data:/data ghcr.io/basvanstein/gsareport:main -p /data/problem.json -d /data -o /output
```
