### 2022-09-29 Basvanstein/GSAreport#18 @rreinecke
> For Linux that should be

Well that did work! I don't know why this did not work for the demo mode then.
