### 2022-09-29 #4721 @rreinecke
> Ok, I can add unit tests for those functions that do not generate files. It will take a bit of time though.

Thanks! As soon as this is implemented I am ready to give my ok. No other remarks.
