+++ src/GSAreport.py
@@ -70,9 +70,9 @@ from bokeh.plotting import figure, output_file, save
 from sklearn.ensemble import RandomForestRegressor
 from sklearn.model_selection import cross_val_score
 
-import plotting.data_processing as dp
-import plotting.interactive_plots as ip
-from plotting.plotting import TS_CODE, make_plot, make_second_order_heatmap
+from .plotting import data_processing as dp
+from .plotting import interactive_plots as ip
+from .plotting.plotting import TS_CODE, make_plot, make_second_order_heatmap
 
 
 class SAReport:
@@ -161,7 +161,7 @@ class SAReport:
         """
         if sample_size > 50 or self.problem["num_vars"] < 64:
             self.x_lhs = latin.sample(
-                self.problem, sample_size * problem["num_vars"], seed=self.seed
+                self.problem, sample_size * self.problem["num_vars"], seed=self.seed
             )
         else:
             self.x_lhs = None
@@ -663,142 +663,142 @@ class SAReport:
             )
         return ([script1, script2], [div1, div2])
 
+if __name__ == "__main__":
+    import argparse
 
-import argparse
+    parser = argparse.ArgumentParser(
+        prog="GSAreport",
+        formatter_class=argparse.RawDescriptionHelpFormatter,
+        description=textwrap.dedent(
+            """\
+            Generate a global sensitivity analysis report for a given data set or function.
+            Common uses cases:
+            --------------------------------
+            Generate samples for evaluation by a real world function / simulator
+                > python GSAreport.py -p problem.json -d data_dir --sample --samplesize 1000
+            Analyse the samples with their output stored in the data folder
+                > python GSAreport.py -p problem.json -d data_dir -o output_dir
+            Analyse a real-world data set and use a Random Forest model to interpolate (data_dir contains x.csv and y.csv)
+                > python GSAreport.py -p problem.json -d data_dir -o output_dir --samplesize 10000
+            """
+        ),
+    )
+    parser.add_argument(
+        "--problem",
+        "-p",
+        default="problem.json",
+        type=str,
+        help="File path to the problem definition in json format.",
+    )
+    parser.add_argument(
+        "--data",
+        "-d",
+        type=str,
+        default="/data/",
+        help="Directory where the intermediate data is stored. (defaults to /data)",
+    )  # will be accesible under args.data
+    parser.add_argument(
+        "--output",
+        "-o",
+        type=str,
+        default="/output/",
+        help="Directory where the output report is stored. (defaults to /output/)",
+    )  # will be accesible under args.output
+    parser.add_argument(
+        "--name",
+        type=str,
+        default="SA",
+        help="Name of the experiment, will be used in the report output.",
+    )
+    parser.add_argument(
+        "--top",
+        type=int,
+        default=10,
+        help="The number of important features to focus on, default is 10.",
+    )
+    parser.add_argument(
+        "--sample",
+        action="store_true",
+        help="When you use this flag, only the samples are generated to be used in the analyse phase.",
+    )
+    parser.add_argument(
+        "--samplesize", "-n", type=int, help="Number of samples to generate.", default=1000
+    )
+    parser.add_argument(
+        "--modelsize", type=int, help="Number of samples for the model.", default=1000
+    )
+    parser.add_argument(
+        "--demo",
+        help="Demo mode, uses a BBOB function as test function.",
+        action="store_true",
+    )
+    args = parser.parse_args()
 
-parser = argparse.ArgumentParser(
-    prog="GSAreport",
-    formatter_class=argparse.RawDescriptionHelpFormatter,
[…]
