+++ src/requirements.txt
@@ -77,7 +77,7 @@ ptyprocess==0.7.0
 pycairo
 pycparser==2.21
 Pygments==2.11.1
-PyGObject
+#PyGObject
 pyOpenSSL==21.0.0
 pyparsing==2.4.7
 pyrsistent==0.18.0
