### 2022-09-29 57e37ff
Added model sampling test
