+++ src/tests/test_model.py
@@ -8,15 +8,15 @@ report = None
 problem = {}
 
 def test_train_model():
-    with open("data/problem.json") as json_file:
+    with open("../data/problem.json") as json_file:
         problem = json.load(json_file)
     fun, opt = bn.instantiate(5, iinstance=1)
     report = GSAreport.SAReport(
         problem,
         top=10,
         name="F5",
-        output_dir="output",
-        data_dir="data",
+        output_dir="../output",
+        data_dir="../data",
         model_samples=128,
     )
     X_lhs, X_morris, X_sobol = report.generateSamples(128)
