+++ src/requirements.txt
@@ -78,6 +78,7 @@ pycairo
 pycparser==2.21
 Pygments==2.11.1
 #PyGObject
+pytest==7.1.3
 pyOpenSSL==21.0.0
 pyparsing==2.4.7
 pyrsistent==0.18.0
