+++ .github/workflows/linter.yml
@@ -14,3 +14,9 @@ jobs:
         run: pip install black==22.3.0
       - name: Run black
         run: black --check --verbose ./src/
+      - name: Install dependencies
+        run: pip install -r ./src/requirements.txt
+      - name: Run pytest
+        run: |
+          cd src
+          pytest
