+++ src/tests/test_loaddata.py
@@ -1,10 +1,13 @@
-import pytest
-from src import GSAreport
 import json
+
+import pytest
+
+from src import GSAreport
 from src.benchmark import bbobbenchmarks as bn
 
 report = None
 problem = {}
+
 
 def test_load_data():
     with open("../data/problem.json") as json_file:
@@ -18,6 +21,12 @@ def test_load_data():
         model_samples=128,
     )
     report.loadData()
-    assert len(report.x_lhs) == 200*problem['num_vars'], f"Unexpected number of samples for LHS {len(report.x_lhs)}"
-    assert len(report.x_morris) == 200*(problem['num_vars']+1), f"Unexpected number of samples for Morris {len(report.x_morris)}"
-    assert len(report.x_sobol) == 200*(2*problem['num_vars']+2), f"Unexpected number of samples for Sobol {len(report.x_sobol)} != {200*(2*problem['num_vars']+2)}"
+    assert (
+        len(report.x_lhs) == 200 * problem["num_vars"]
+    ), f"Unexpected number of samples for LHS {len(report.x_lhs)}"
+    assert len(report.x_morris) == 200 * (
+        problem["num_vars"] + 1
+    ), f"Unexpected number of samples for Morris {len(report.x_morris)}"
+    assert len(report.x_sobol) == 200 * (
+        2 * problem["num_vars"] + 2
+    ), f"Unexpected number of samples for Sobol {len(report.x_sobol)} != {200*(2*problem['num_vars']+2)}"
