### 2022-09-29 330cd09
Running pytest via python works
