### 2022-09-29 #4721 @Basvanstein
> > Ok, I can add unit tests for those functions that do not generate files. It will take a bit of time though.
>
> Thanks! As soon as this is implemented I am ready to give my ok. No other remarks.

Thank you! I have added pytest tests in the src/test folder and have updated the readme to include instructions. Due to relative imports you need to run

```
cd src
python -m pytest
```

in order to execute the tests. I execute them in the automated github action as well on each merge request.
The tests cover data loading, data generation, generation of the bokeh plots, sampling with default scheme, sampling with model and model training.
