### 2022-09-29 Basvanstein/GSAreport#20 @xtruan
Update README Docker commands to handle cases where working directory has spaces
The Docker commands will fail if the users local working directory has a space. I recommend changing all instances of:

`` `pwd` `` to `"$(pwd)"`

In addition, specifying docker volumes on their own lines will also present issues. I recommend changing the following command from:
```zsh
docker run --rm -v `pwd`/output:/output \
    -v `pwd`/data:/data \
    ghcr.io/basvanstein/gsareport:main -p /data/problem.json -d /data -o /output
```
to:
```zsh
docker run --rm -v "$(pwd)"/output:/output -v "$(pwd)"/data:/data \
    ghcr.io/basvanstein/gsareport:main -p /data/problem.json -d /data -o /output
```
