### 2022-09-29 #4721 @xtruan
Thank you for addressing my earlier issues and for making the other fixes outlined. I was able to build and test the application from end to end. I very much like the look of the generated reports!

I have created one last minor issue to deal with spaces in the user's working directory impacting the Docker commands. When this is resolved I will check my final box and give my ok: https://github.com/Basvanstein/GSAreport/issues/20
