### 2022-09-30 #4721 @Basvanstein
> Thank you for addressing my earlier issues and for making the other fixes outlined. I was able to build and test the application from end to end. I very much like the look of the generated reports!
>
> I have created one last minor issue to deal with spaces in the user's working directory impacting the Docker commands. When this is resolved I will check my final box and give my ok: [Basvanstein/GSAreport#20](https://github.com/Basvanstein/GSAreport/issues/20)

Thank you for your review and the positive comments! Good point about the `pwd` and docker instructions. I have updated the readme and online documentation accordingly.
