+++ src/requirements.txt
@@ -55,7 +55,6 @@ nbformat==5.1.3
 nest-asyncio==1.5.4
 notebook==6.4.4
 numba==0.56.2
-numpy==1.23.3
 ordered-set==4.0.2
 packaging==21.3
 pandas==1.3.5
