+++ paper.bib
@@ -32,3 +32,13 @@
   year={2015},
   publisher={Elsevier}
 }
+@ARTICLE{9903639,  
+    author={Stein, Bas Van and Raponi, Elena and Sadeghi, Zahra and Bouman, Niek and Van Ham, Roeland C. H. J. and Bäck, Thomas},  
+    journal={IEEE Access},   
+    title={A Comparison of Global Sensitivity Analysis Methods for Explainable AI With an Application in Genomic Prediction},   
+    year={2022},  
+    volume={10},  
+    number={},  
+    pages={103364-103381},  
+    doi={10.1109/ACCESS.2022.3210175}
+}
