+++ paper.bib
@@ -1,5 +1,5 @@
 
-@article{SALibHerman2017,
+@article{Herman:2017,
     doi = {10.21105/joss.00097},
     url = {https://doi.org/10.21105/joss.00097},
     year  = {2017},
@@ -7,12 +7,12 @@
     publisher = {The Open Journal},
     volume = {2},
     number = {9},
-    author = {Jon Herman and Will Usher},
+    author = {Herman, Jon and Usher, Will},
     title = {{SALib}: An open-source Python library for Sensitivity Analysis},
     journal = {The Journal of Open Source Software}
 }
 
-@incollection{shapNIPS2017_7062,
+@incollection{Lundberg:2017,
     title = {A Unified Approach to Interpreting Model Predictions},
     author = {Lundberg, Scott M and Lee, Su-In},
     booktitle = {Advances in Neural Information Processing Systems 30},
@@ -23,7 +23,7 @@
     url = {http://papers.nips.cc/paper/7062-a-unified-approach-to-interpreting-model-predictions.pdf}
 }
 
-@article{pianosi2015matlab,
+@article{Pianosi:2015,
   title={A Matlab toolbox for global sensitivity analysis},
   author={Pianosi, Francesca and Sarrazin, Fanny and Wagener, Thorsten},
   journal={Environmental Modelling \& Software},
@@ -32,7 +32,7 @@
   year={2015},
   publisher={Elsevier}
 }
-@ARTICLE{9903639,  
+@ARTICLE{Stein:2022,  
     author={Stein, Bas Van and Raponi, Elena and Sadeghi, Zahra and Bouman, Niek and Van Ham, Roeland C. H. J. and Bäck, Thomas},  
     journal={IEEE Access},   
     title={A Comparison of Global Sensitivity Analysis Methods for Explainable AI With an Application in Genomic Prediction},   
