### 2022-10-13 #4721 @kthyng
Hi @Basvanstein! Would you like to make a new release that is associated with the JOSS paper? If so what is the new version number?
