### 2022-10-13 #4721 @Basvanstein
Yes the release is made. Please set the version to v1.0.1 such that it is the same as the release version tag.
Thank you! Will now change the ending parenthesis.
