### 2022-10-13 #4721 @Basvanstein
Fixed the parenthesis in the `paper.md` file.
